Issue #43: Move doc definition loader/validation components to src directory
The `etc` directory has become a bit of a dumping ground for all JavaScript and etc. files. Start to clean it up by moving newly-introduced JavaScript components into a new `src` directory.
