+++ etc/sync-function-loader.js
@@ -10,8 +10,8 @@ exports.load = loadFromFile;
 var fs = require('fs');
 var path = require('path');
 var indent = require('../lib/indent.js/indent.min.js');
-var docDefinitionsLoader = require('./document-definitions-loader.js');
-var fileFragmentLoader = require('./file-fragment-loader.js');
+var docDefinitionsLoader = require('../src/document-definitions-loader.js');
+var fileFragmentLoader = require('../src/file-fragment-loader.js');
 
 function loadFromFile(docDefinitionsFile) {
   var syncFuncTemplateDir = path.dirname(module.filename);
