+++ src/document-definitions-validator.js
