+++ src/file-fragment-loader.js
